Implement copy version: http://bugzilla.redhat.com/show_bug.cgi?id=1104015 Squashed commit of the following:
commit ca25839b932ee9ee962099c47b8582f2a6731b41
Merge: 30a2fbe 87695aa
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 30 15:24:37 2014 +1000

    Merge branch 'integration/master' into rhbz1104015

    Conflicts:
    	zanata-war/src/main/resources/messages.properties

commit 30a2fbe19100e7fa3aaf5469b5a658924cef3e6a
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 30 15:12:35 2014 +1000

    Refactor JPACopier

commit 53e441459ae9c6239acc10f18950d80dd062e2ad
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 30 08:53:54 2014 +1000

    WIP: fixing test

commit 639efdc86068bd4695c8d729a0cefd83783981de
Merge: dfc12f3 2e0439b
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 29 15:44:25 2014 +1000

    Merge branch 'integration/master' into rhbz1104015

commit dfc12f33c67f525ac44360301443f71b06eb67ff
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 29 15:11:44 2014 +1000

    WIP: refactoring copy bean implementation

commit e85c7c19b1182f3a7afc77ea82c8f634f8197c60
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 29 08:29:55 2014 +1000

    Refactor return type

commit 509e7f98bd52c72fd9895db5a93e0c24ea290992
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 28 09:24:36 2014 +1000

    Fix test

commit 885727c0839ffdf1084c5a118c2b51b7acd4fc73
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 28 07:41:57 2014 +1000

    Refactor beanUtil

commit 028336ae8ec9544f1a6cce9368a70f9da5c85345
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 23 15:05:56 2014 +1000

    Update comments, refactor copy of Collection type in beanUtils

commit aa3ea50dccd445940f60079ee1294b55e83324f4
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 23 10:03:40 2014 +1000

    Update powermock version used in zanata-model

commit 8ada818fed2b180f8941910716faf25c83d5e3f6
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 22 12:42:54 2014 +1000

    Fix plural copy, move Copy annotation to getter method

commit 5dcfd0ec06179b7bef9476901fe760623f2c2c39
Merge: f9710b4 4f02f0a
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 21 10:40:09 2014 +1000

    Merge branch 'integration/master' into rhbz1104015

commit f9710b467f211052fd30a06ac62d2fed420a7e51
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 21 09:42:31 2014 +1000

    Add unit test for beanUtils

commit ea5d7804da100b5b9f590c15eb9908fd2ebb4613
Author: Alex Eng <aeng@redhat.com>
Date:   Fri Jul 18 10:30:30 2014 +1000

    Implement annotation Copy for cloning property

commit be178d518dbc29ca9e08e6cf5932d02bfd1e3a34
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 17 15:08:54 2014 +1000

    WIP: copy version refactoring

commit 6e07eed21a1acbac8c77c85a479bf6bb41def14f
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 17 10:16:58 2014 +1000

    WIP: Copy annotation

commit 0d80531f850d387449913014c09563e3f93c6480
Author: Alex Eng <aeng@redhat.com>
Date:   Fri Jul 11 15:05:09 2014 +1000

    Implement copy version progress in version-tab

commit 80d8404f632a622367dcd25b85e4debbe6b4f8b5
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 10 15:24:31 2014 +1000

    WIP: show copy progress in project page

commit d43624b57e92f30af4098dd26639a00e48f99bdf
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 9 15:16:02 2014 +1000

    add batch process unit test

commit 6a2f3fa47370c4532d0d7a3101a6e8d530445d7f
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 8 15:20:49 2014 +1000

    Round to 4 decimal for progress bar complete

commit b029e3c20418b02abab3dca3d899a88336bfe4d6
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 8 15:10:06 2014 +1000

    Implement copy version: http://bugzilla.redhat.com/show_bug.cgi?id=1104015

+++ zanata-war/src/main/java/org/zanata/ui/ProgressBar.java
@@ -29,5 +29,5 @@ public interface ProgressBar {
 
     boolean isInProgress();
 
-    int getCompletedPercentage();
+    double getCompletedPercentage();
 }
